This seemed like a simple solution, at least for the time being.
