+++ src/version.h
@@ -21,8 +21,8 @@
 #define __XDMA_VERSION_H__
 
 #define DRV_MOD_MAJOR		2020
-#define DRV_MOD_MINOR		0
-#define DRV_MOD_PATCHLEVEL 0
+#define DRV_MOD_MINOR		1
+#define DRV_MOD_PATCHLEVEL	0
 
 #define DRV_MODULE_VERSION      \
 	__stringify(DRV_MOD_MAJOR) "." \
